delivery
finishing delivery form and about to start js validation

+++ app/Http/Controllers/DeliveriesController.php
@@ -5,9 +5,11 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
 use App\Delivery;
+use App\Http\Requests\Delivery;
 
 class DeliveriesController extends Controller
 {
+
     /**
      * Display a listing of the resource.
      *
@@ -24,9 +26,13 @@ class DeliveriesController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function create()
+    public function create($userId)
     {
-        //llamar a vista que crea deliveries
+		/*if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}*/
+
+        return view('partials.sportForm', array('userId'=>$userId));
     }
 
     /**
@@ -35,8 +41,12 @@ class DeliveriesController extends Controller
      * @param  \Illuminate\Http\Request  $request
      * @return \Illuminate\Http\Response
      */
-    public function store(Delivery $request)
+    public function store(Delivery $request, $deliveryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
         $delivery = new Delivery();
 
 		$delivery->name = $request->input('name');
@@ -44,6 +54,7 @@ class DeliveriesController extends Controller
 		$delivery->price = $request->input('price');
 
 		$delivery->save();
+		return redirect("user/$userId/deliveries");
     }
 
     /**
@@ -52,12 +63,13 @@ class DeliveriesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function edit($deliveryId)
+    public function edit($userId, $deliveryId)
     {
 		if (!$delivery = Delivery::find($deliveryId)) {
 			abort(404);
 		}
-		//llamar vista con datos de delivery
+
+		return view('partials.deliveryForm', array('userId'=>$userId, 'delivery'=>$delivery));
     }
 
     /**
@@ -67,8 +79,12 @@ class DeliveriesController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function update(Delivery $request, $deliveryId)
+    public function update(Delivery $request, $userId, $deliveryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$delivery = Delivery::find($deliveryId)) {
 			abort(404);
 		}
@@ -78,6 +94,7 @@ class DeliveriesController extends Controller
 		$delivery->price = $request->input('price');
 
 		$delivery->save();
+		return redirect("user/$userId/deiveries");
     }
 
     /**
@@ -88,10 +105,16 @@ class DeliveriesController extends Controller
      */
     public function destroy($deliveryId)
     {
+		if(!Auth::check() || Auth::user()->role!="root") {
+			abort(404);
+		}
+
 		if (!$delivery = Delivery::find($deliveryId)) {
 			abort(404);
 		}
 
 		$delivery->delete();
+		return redirect("user/$userId/deiveries");
     }
+
 }
